Tests/LibWeb/Text/expected/Worker/SharedWorker-type-mismatch.txt

worker1: connect 1
worker2: error
